Update app.js
I was having trouble displaying negative electrical usage (associated with solar generation) for SCE data, and the problem was the graph drawing function munging the display when the usage data was negative and the yMin*0.95 reduced the bounds of the graph such that the data extended beyond the bounds... though strangely only if the negative data occurred within multiple days (a single data plotted just fine).  

This fix eliminates the errors associated with assuming that the usage data (max or min) are always positive or negative, and so the resulting grapher should work with a broader range of SCE customer data, and particularly better for customers with solar panels or other on-side power generation.